% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{AICc}
\alias{AICc}
\title{Corrected Akaike Information Criterion}
\usage{
AICc(model)
}
\arguments{
\item{model}{Some object that defines the \code{\link[=logLik]{logLik()}} function}
}
\description{
I think this is Williamson's correction, but of course I will look up the
proper attribution before I publish this too widely. The idea is that extra
parameters are penalized more harshly when there is a small sample size.
}
