% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.update.R
\name{recurrent.update}
\alias{recurrent.update}
\title{Build a List of Formulae from Combinations of Terms}
\usage{
recurrent.update(form, extra.terms)
}
\arguments{
\item{form}{The formula that will get extra terms}

\item{extra.terms}{The additional terms that will be added with tail
recursion.}
}
\value{
A list of formulae that includes every combination of \code{model} and
each item in \code{extra.terms}.
}
\description{
If you want to compare a bunch of models to select the best one then it is
helpful to quickly generate every possible model that combines your terms.
}
