% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{data}
\name{results}
\alias{results}
\title{Spectral slopes from simulated power law noise.}
\format{
A data frame with 1440768 rows and 4 variables:
\describe{
\item{Power:}{The expected power law, \eqn{\alpha}, for the simulation.}
\item{Degree:}{The number of terms in the AR model, \eqn{K}.}
\item{Size:}{The number of samples in the time series, \eqn{N}.}
\item{Slope:}{The observed slope of time series' log-log power spectrum.}
\item{Positivized.Error:}{The deviation of the slope from the power, but
positive.}
\item{Degree.Ratio}{The Degree divided by the square root of the Size.}
}
}
\usage{
results
}
\description{
This data set comes from running the default simulation in pypowerlawnoise.
There are spectral slopes from many different time series generated with
Kasdin's autoregressive (AR) model.
}
\keyword{datasets}
